refactor(http): isolate file-backed body implementation
